requirements.txt
# Cœur réseau et API
scapy
requests
fastapi
uvicorn
pydantic
pysnmp
pycryptodomex

# Interface graphique
flet

# Alertes SMS
twilio
python-dotenv

# Outils de test et qualité (pour la pipeline)
pytest
httpx
pylint
